Some WIP USB serial shenanigans. Probably don't treat this code as working just yet

+++ RP23V50firmware/src/RotaryEncoder.cpp
@@ -135,8 +135,8 @@ int lastRotaryDivider = 8;
 int rotaryDivider = 8;
 
 bool resetEncoderPosition = false;
-int32_t encoderPositionOffset = 0;
-volatile int32_t encoderPosition = 0;
+long encoderPositionOffset = 0;
+volatile long encoderPosition = 0;
 
 volatile int numberOfSteps = 0;
